Fixes getClientId for ias tokens, implement getAudiences()/getClientId according to OIDC Connect spec
