+++ spring-xsuaa/src/main/java/com/sap/cloud/security/xsuaa/extractor/IasToken.java
@@ -14,10 +14,8 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.security.Principal;
 import java.time.Instant;
-import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Objects;
-import java.util.Set;
 
 import static com.sap.cloud.security.token.TokenClaims.*;
 
@@ -42,7 +40,7 @@ class IasToken implements Token {
 
 	@Override
 	public boolean hasClaim(@Nonnull String claimName) {
-		return decodedToken.containsClaim(claimName);
+		return decodedToken.hasClaim(claimName);
 	}
 
 	@Nullable
@@ -96,17 +94,7 @@ class IasToken implements Token {
 	}
 
 	@Override
-	public Set<String> getAudiences() {
-		return new LinkedHashSet<>(getClaimAsStringList(AUDIENCE));
-	}
-
-	@Override
 	public String getZoneId() {
 		return decodedToken.getClaimAsString(SAP_GLOBAL_ZONE_ID);
-	}
-
-	@Override
-	public String getClientId() {
-		return decodedToken.getClaimAsString(SAP_GLOBAL_USER_ID);
 	}
 }
